+++ src/tdastro/effects/effect_model.py
@@ -1,0 +1,47 @@
+"""The base EffectModel class used for all effects."""
+
+from tdastro.base_models import ParameterizedNode
+
+
+class EffectModel(ParameterizedNode):
+    """A physical or systematic effect to apply to an observation."""
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def __str__(self):
+        """Return the string representation of the model."""
+        return "EffectModel"
+
+    def required_parameters(self):
+        """Returns a list of the parameters of a PhysicalModel
+        that this effect needs to access.
+
+        Returns
+        -------
+        parameters : `list` of `str`
+            A list of every required parameter the effect needs.
+        """
+        return []
+
+    def apply(self, flux_density, wavelengths=None, physical_model=None, **kwargs):
+        """Apply the effect to observations (flux_density values)
+
+        Parameters
+        ----------
+        flux_density : `numpy.ndarray`
+            A length T X N matrix of flux density values.
+        wavelengths : `numpy.ndarray`, optional
+            A length N array of wavelengths.
+        physical_model : `PhysicalModel`
+            A PhysicalModel from which the effect may query parameters
+            such as redshift, position, or distance.
+        **kwargs : `dict`, optional
+           Any additional keyword arguments.
+
+        Returns
+        -------
+        flux_density : `numpy.ndarray`
+            A length T x N matrix of flux densities after the effect is applied.
+        """
+        raise NotImplementedError()
