+++ tests/tdastro/test_base_models.py
@@ -1,8 +1,7 @@
 import random
 
 import pytest
-from astropy.cosmology import WMAP9
-from tdastro.base_models import ParameterizedModel, PhysicalModel
+from tdastro.base_models import ParameterizedNode
 
 
 def _sampler_fun(**kwargs):
@@ -16,7 +15,7 @@ def _sampler_fun(**kwargs):
     return random.random()
 
 
-class PairModel(ParameterizedModel):
+class PairModel(ParameterizedNode):
     """A test class for the parameterized model.
 
     Attributes
@@ -34,9 +33,9 @@ class PairModel(ParameterizedModel):
 
         Parameters
         ----------
-        value1 : `float`, `function`, `ParameterizedModel`, or `None`
+        value1 : `float`, `function`, `ParameterizedNode`, or `None`
             The first value.
-        value2 : `float`, `function`, `ParameterizedModel`, or `None`
+        value2 : `float`, `function`, `ParameterizedNode`, or `None`
             The second value.
         **kwargs : `dict`, optional
            Any additional keyword arguments.
@@ -142,31 +141,3 @@ def test_parameterized_model_modify() -> None:
     # We cannot set a value that hasn't been added.
     with pytest.raises(KeyError):
         model.set_parameter("brightness", 5.0)
-
-
-def test_physical_model():
-    """Test that we can create a physical model."""
-    # Everything is specified.
-    model1 = PhysicalModel(ra=1.0, dec=2.0, distance=3.0, redshift=0.0)
-    assert model1.ra == 1.0
-    assert model1.dec == 2.0
-    assert model1.distance == 3.0
-    assert model1.redshift == 0.0
-
-    # Derive the distance from the redshift using the example from:
-    # https://docs.astropy.org/en/stable/api/astropy.cosmology.units.redshift_distance.html
-    model2 = PhysicalModel(ra=1.0, dec=2.0, redshift=1100.0, cosmology=WMAP9)
-    assert model2.ra == 1.0
-    assert model2.dec == 2.0
-    assert model2.redshift == 1100.0
-    assert model2.distance == pytest.approx(14004.03157418 * 1e6)
-
-    # Neither distance nor redshift are specified.
-    model3 = PhysicalModel(ra=1.0, dec=2.0)
-    assert model3.redshift is None
-    assert model3.distance is None
-
-    # Redshift is specified but cosmology is not.
-    model4 = PhysicalModel(ra=1.0, dec=2.0, redshift=1100.0)
-    assert model4.redshift == 1100.0
-    assert model4.distance is None
